+++ Sender/tst/UT/BMSDataSender/Mocks.h
@@ -3,7 +3,7 @@
 
 extern char Printf_FormartString[100];
 extern int call_Printf;
-extern double printf_floatpar_data[2];
+extern float printf_floatpar_data[2];
 extern double printf_floatpar_AlldataSamples[4][2];
 extern int numberofcallsToRequestStop;
 extern int print_Mock_FilenotfoundInstance(char *Par_s);
